Checkpoint: Corrigido erro crítico de deployment que impedia a inicialização do servidor:
**Problema:**
- ReferenceError: tinyint is not defined em drizzle/schema.ts linha 130
- Campo `isResolved` em `contract_alerts` estava usando `tinyint()` que não é uma função válida do Drizzle

**Solução:**
- Substituído `tinyint()` por `int()` que é o tipo correto do Drizzle para MySQL
- Mantida a lógica de default(0) para representar booleano como número

**Resultado:**
- ✅ Servidor inicia corretamente
- ✅ Projeto pronto para deployment
- ✅ Sem erros de referência de tipos

+++ drizzle/schema.ts
@@ -127,7 +127,7 @@ export const contractAlerts = mysqlTable("contract_alerts", {
 	spaceId: int().notNull().references(() => consumableSpaces.id),
 	alertType: mysqlEnum(['monthly_payment','contract_expiry']).notNull(),
 	daysUntilEvent: int().notNull(),
-	isResolved: tinyint().default(0).notNull(),
+	isResolved: int().default(0).notNull(),
 	createdAt: timestamp({ mode: 'string' }).default('CURRENT_TIMESTAMP').notNull(),
 	resolvedAt: timestamp({ mode: 'string' }),
 });
